+++ src/super/i-static-page/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :house: Internal
+
+* Removed `jasmine` deps in the `runtime.debug` mode
+
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :bug: Bug Fix
+
+* Provided the `lang` attribute to `<html>`
+
 ## v3.0.0-rc.169 (2021-03-25)
 
 #### :bug: Bug Fix